backend/requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
python-dotenv
python-jose
passlib
psycopg2-binary
pydantic[email]